# server/docs/swagger.yaml
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.6
paths:
...
  /fit/convert_to_csv:
    post:
      consumes:
      - multipart/form-data
      description: 上传FIT文件并将其内容转换为CSV格式返回
      parameters:
      - description: FIT运动数据文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - text/csv
      responses:
        "200":
          description: 转换成功，返回CSV文件
          schema:
            type: file
        "400":
          description: 转换失败
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 将FIT文件转换为CSV格式
      tags:
      - FIT
  /fit/session/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定会话的完整数据,包括Lap和Record
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取单个运动会话详情
      tags:
      - FIT
  /fit/sessions:
    get:
      consumes:
      - application/json
      description: 分页获取用户的运动会话记录
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取运动会话列表
      tags:
      - FIT
  /fit/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传FIT文件并解析其中的运动数据(Session/Lap/Record)
      parameters:
      - description: FIT运动数据文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
        "400":
          description: 上传失败
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 上传FIT文件
      tags:
      - FIT
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...